+++ R/get_transcripts.R
@@ -19,39 +19,51 @@
 #' library(TxDb.Hsapiens.UCSC.hg19.knownGene)
 #' txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene
 #' GenomicFeatures::genes(txdb)
-#' }
+#' } 
 #' @inheritParams plot_locus
+#' @inheritParams get_tx_biotypes
 #' @keywords internal
 #' @importFrom GenomicRanges seqnames findOverlaps makeGRangesFromDataFrame
 #' @importFrom dplyr arrange desc filter row_number ungroup n_distinct
 #' @importFrom data.table as.data.table
+#' @importFrom ensembldb transcripts listColumns
+#' @importFrom AnnotationFilter AnnotationFilterList SeqNameFilter TxIdFilter
+#' @importFrom S4Vectors subjectHits
 get_transcripts <- function(gr.snp,
                             max_transcripts=1,
                             remove_pseudogenes=TRUE,
+                            tx_biotypes=NULL,
                             verbose=TRUE){
     symbol <- tx_name <- tx_biotype <- width <- NULL;
-    requireNamespace("EnsDb.Hsapiens.v75") 
-    requireNamespace("AnnotationFilter")
-    requireNamespace("ensembldb")
-    requireNamespace("S4Vectors")
+    requireNamespace("EnsDb.Hsapiens.v75")  
     
+    ##### Check transcript biotypes #####
+    if(!is.null(tx_biotypes)){
+        tx_filter <- get_tx_biotypes(tx_biotypes=tx_biotypes,
+                                     verbose=verbose)
+    } else {
+        tx_filter <- NULL
+    }
     txdb <- EnsDb.Hsapiens.v75::EnsDb.Hsapiens.v75
     #### Query genome db with finemap_DT coordinates ####
     txdb_transcripts <- ensembldb::transcripts(
         txdb,
-      columns = c(ensembldb::listColumns(txdb , "tx"), "symbol"),
+      columns = c(ensembldb::listColumns(txdb ,"tx"),"symbol"),
       filter = AnnotationFilter::AnnotationFilterList(
-          # GRangesFilter(gr.snp)
+          #### Filter by seqnames ####
           AnnotationFilter::SeqNameFilter(
-              value = unique(GenomicRanges::seqnames(gr.snp)) )
-                                              ))
+              value = unique(GenomicRanges::seqnames(gr.snp))
+              ),
+          #### Filter by tx_biotype ####
+          tx_filter
+          )
+      ) 
     #### Find exact overlap between transcripts and finemap_DT ####
     # PROBLEM: This
     hits <- GenomicRanges::findOverlaps(query = gr.snp,
                                         subject = txdb_transcripts,
                                         type="any",
                                         ignore.strand=TRUE)
-    
     #### limit the number of transcript per gene ####
     tx.filt <- data.frame(txdb_transcripts[S4Vectors::subjectHits(hits),]) |>
         ## !!IMPORTANT!! If unique() isn't applied here,
@@ -65,9 +77,8 @@ get_transcripts <- function(gr.snp,
         dplyr::filter(dplyr::row_number() %in% seq_len(max_transcripts)) |>
         dplyr::ungroup() |> 
         data.table::as.data.table()
-    
     #### remove_pseudogenes ####
-    if(remove_pseudogenes){
+    if(isTRUE(remove_pseudogenes)){
         pseudo <- grep("pseudogene|nonsense_mediated_decay",
                        unique(tx.filt$tx_biotype),value = TRUE)
         tx.filt <- subset(tx.filt,
@@ -76,7 +87,7 @@ get_transcripts <- function(gr.snp,
                               (!startsWith(symbol,"RPS"))
         )
     }
-    #### Convert back to Granges ####
+    #### Convert back to Granges #### 
     tx.gr <- GenomicRanges::makeGRangesFromDataFrame(
         df = tx.filt, 
         seqnames.field = "seqnames",