+++ R/guess_pvalue_col.R
@@ -1,10 +1,11 @@
 guess_pvalue_col <- function(dat,
-                             QTL_prefix=NULL){
+                             qtl_suffix=NULL){
     requireNamespace("stringr")
+    
     p_options <- c("p","p-value","p-values","pvalue","pvalues","pval")
     p_options <- c(p_options, stringr::str_to_sentence(p_options))
-    pval_col <- paste0(QTL_prefix,p_options)[
-        paste0(QTL_prefix,p_options) %in% colnames(dat)
-    ][1]
+    pval_col <- paste0(p_options,qtl_suffix)[
+        paste0(p_options,qtl_suffix) %in% colnames(dat)
+    ] 
     return(pval_col)
 }