+++ R/snp_track_merged.R
@@ -5,12 +5,14 @@
 #' @keywords internal
 #' @inheritParams plot_locus
 #' @importFrom echodata melt_finemapping_results
+#' @importFrom echoannot add_mb
 #' @importFrom stats as.formula
 snp_track_merged <- function(dat,
                               yvar="-log10(P)",
                               labels_subset = c("Lead","UCS","Consensus"),
                               absolute_labels=FALSE,
                               label_type="rsid_only",
+                              label_leadsnp=TRUE,
                               sig_cutoff=5e-8,
                               cutoff_lab=paste("P <",sig_cutoff),
                               point_alpha=.5,
@@ -22,8 +24,6 @@ snp_track_merged <- function(dat,
                               strip.text.y.angle=0,
                               show_plot=FALSE,
                               verbose=TRUE){
-    # labels_subset = c("Lead","UCS","Consensus"); yvar="-log10(P)"; absolute_labels=F; label_type="rsid_only";  sig_cutoff=5e-8;
-    # cutoff_lab=paste("P <",sig_cutoff); point_alpha=.5; show.legend=T; xtext=T;  facet_formula="Method~."; track_heights=NULL;
     requireNamespace("ggplot2")
     Mb <- NULL;
     
@@ -46,7 +46,7 @@ snp_track_merged <- function(dat,
         melt_methods <- FALSE
         grouping_vars <- c("SNP")
     }
-    finemap_melt$Mb <- finemap_melt$POS/1000000
+    finemap_melt <- echoannot::add_mb(dat = finemap_melt)
     
     # Plot
     snp_plot <- ggplot2::ggplot(
@@ -58,7 +58,8 @@ snp_track_merged <- function(dat,
         ggplot2::scale_color_gradient(low="blue",high ="red",
                              breaks=c(0,.5,1), limits=c(0,1)) +
         ## Sig cutoff line
-        ggplot2::geom_hline(yintercept=sig_cutoff, alpha=.5, linetype=2, size=.5, 
+        ggplot2::geom_hline(yintercept=sig_cutoff, 
+                            alpha=.5, linetype=2, size=.5, 
                    color="black") +
         ggplot2::geom_text(ggplot2::aes(x=min(Mb), y=sig_cutoff*1.1),
                   label=cutoff_lab,
@@ -75,7 +76,8 @@ snp_track_merged <- function(dat,
             strip.text.y = ggplot2::element_text(angle=strip.text.y.angle)
         ) +
         ggplot2::guides(
-            fill = ggplot2::guide_colourbar(barwidth = 1, barheight = 3))
+            fill = ggplot2::guide_colourbar(barwidth = 1,
+                                            barheight = 3))
     
     #  Choose breaks and ylims
     if(startsWith(yvar,"-log")){
@@ -110,6 +112,15 @@ snp_track_merged <- function(dat,
         snp_plot <- snp_plot +
             ggplot2::theme(axis.text.x = ggplot2::element_blank(),
                   axis.title.x = ggplot2::element_blank())
+    }
+    if(isTRUE(label_leadsnp)){
+        snp_plot <- snp_plot + 
+            ggplot2::geom_point(data = subset(finemap_melt, leadSNP),
+                                ggplot2::aes_string(x=genomic_units,
+                                                    y=yvar),
+                                color="red",pch=9, size=3, 
+                                show.legend = FALSE, alpha=1) +
+            ggplot2::theme(axis.title.x = ggplot2::element_blank())
     }
     if(show_plot) print(snp_plot)
     return(snp_plot)